refactor: comment 패키지 생성 및 의존성 분리
