+++ backend/src/main/java/com/digginroom/digginroom/comment/repository/dto/CommentMember.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.repository.dto;
+package com.digginroom.digginroom.comment.repository.dto;
 
 import java.time.LocalDateTime;
