+++ backend/src/main/java/com/digginroom/digginroom/comment/service/dto/CommentsResponse.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.service.dto;
+package com.digginroom.digginroom.comment.service.dto;
 
 import java.util.List;
